+++ breeze-core/src/main/kotlin/extensions/BreezeComponentExtensions.kt
@@ -5,12 +5,14 @@ package com.windea.breezeframework.core.extensions
 
 import com.windea.breezeframework.core.annotations.*
 import com.windea.breezeframework.core.component.*
+import java.lang.reflect.*
 import java.nio.charset.*
 
-//region Any Extensions
+//region Converter Extensions
 /**
  * 将当前对象转化为指定类型。如果转换失败，则抛出异常。转化后的对象是基于一般转化逻辑得到的新对象。
  */
+@BreezeComponentExtension
 inline fun <reified T> Any?.convert(): T {
 	return if(this is T) this else Converter.convert(this)
 }
@@ -18,48 +20,13 @@ inline fun <reified T> Any?.convert(): T {
 /**
  * 将当前对象转化为指定类型。如果转换失败，则返回null。转化后的对象是基于一般转化逻辑得到的新对象。
  */
+@BreezeComponentExtension
 inline fun <reified T> Any?.convertOrNull(): T? {
 	return if(this is T) this else Converter.convertOrNull(this)
 }
 //endregion
 
-//region String Extensions
-/**
- * 根据指定的加密器，加密当前字节数组。
- */
-@BreezeComponentExtension
-fun ByteArray.encryptBy(encrypter: Encrypter,secret:ByteArray?=null): ByteArray {
-	return encrypter.encrypt(this,secret)
-}
-
-/**
- * 根据指定的加密器，解密当前字符串。某些加密算法可能不支持解密。
- *
- * @throws UnsupportedOperationException 如果不支持解密。
- */
-@BreezeComponentExtension
-fun ByteArray.decryptBy(encrypter: Encrypter,secret: ByteArray?=null): ByteArray {
-	return encrypter.decrypt(this,secret)
-}
-
-
-/**
- * 根据指定的编码器，编码当前字符串，以指定的字符集显示。
- */
-@BreezeComponentExtension
-fun String.encodeBy(encoder: Encoder, charset: Charset = Charsets.UTF_8): String {
-	return encoder.encode(this, charset)
-}
-
-/**
- * 根据指定的编码器，解码当前字符串，以指定的字符集显示。
- */
-@BreezeComponentExtension
-fun String.decodeBy(encoder: Encoder, charset: Charset = Charsets.UTF_8): String {
-	return encoder.decode(this, charset)
-}
-
-
+//region Escaper Extensions
 /**
  * 根据指定的转义器，转义当前字符串。
  *
@@ -79,8 +46,47 @@ fun String.escapeBy(escaper: Escaper): String {
 fun String.unescapeBy(escaper: Escaper): String {
 	return escaper.unescape(this)
 }
+//endregion
 
+//region Decoder Extensions
+/**
+ * 根据指定的编码器，编码当前字符串，以指定的字符集显示。
+ */
+@BreezeComponentExtension
+fun String.encodeBy(encoder: Encoder, charset: Charset = Charsets.UTF_8): String {
+	return encoder.encode(this, charset)
+}
 
+/**
+ * 根据指定的编码器，解码当前字符串，以指定的字符集显示。
+ */
+@BreezeComponentExtension
+fun String.decodeBy(encoder: Encoder, charset: Charset = Charsets.UTF_8): String {
+	return encoder.decode(this, charset)
+}
+//endregion
+
+//region Encrypter Extensions
+/**
+ * 根据指定的加密器，加密当前字节数组。
+ */
+@BreezeComponentExtension
+fun ByteArray.encryptBy(encrypter: Encrypter,secret:ByteArray?=null): ByteArray {
+	return encrypter.encrypt(this,secret)
+}
+
+/**
+ * 根据指定的加密器，解密当前字符串。某些加密算法可能不支持解密。
+ *
+ * @throws UnsupportedOperationException 如果不支持解密。
+ */
+@BreezeComponentExtension
+fun ByteArray.decryptBy(encrypter: Encrypter,secret: ByteArray?=null): ByteArray {
+	return encrypter.decrypt(this,secret)
+}
+//endregion
+
+//region CaseType Extensions
 /**
  * 尝试推断当前字符串的字母格式。
  */
@@ -169,8 +175,9 @@ fun String.switchCaseBy(targetCaseType: CaseType): String {
 	val sourceLetterCase = inferLetterCase() ?: throw IllegalArgumentException("Cannot infer case type for string '$this'.")
 	return splitBy(sourceLetterCase).joinToStringBy(targetCaseType)
 }
+//endregion
 
-
+//region PathType Extensions
 /**
  * 根据指定的路径类型，判断当前字符串是否匹配指定的路径。
  * 默认使用标准路径[PathType.StandardPath]。
@@ -192,9 +199,8 @@ fun String.matchesBy(path: String, pathType: PathType = PathType.StandardPath): 
 fun String.resolveVariablesBy(path: String, pathType: PathType = PathType.StandardPath): Map<String, String> {
 	return pathType.resolveVariables(this, path)
 }
-//endregion
 
-//region Collection Extensions
+
 /**
  * 根据指定路径和指定路径类型查询当前数组，返回查询结果列表。
  * 如果指定路径为空路径，则目标返回查询对象的单例列表。
@@ -340,5 +346,39 @@ fun <T> List<*>.getOrElseBy(path: String, pathType: PathType = PathType.Standard
 @BreezeComponentExtension
 fun <T> Map<*, *>.getOrElseBy(path: String, pathType: PathType = PathType.StandardPath, defaultValue: () -> T): T {
 	return pathType.getOrElse(this, path, defaultValue)
+}
+//endregion
+
+//region DataType Extensions
+/**
+ * 根据指定的数据类型，序列化当前对象。
+ */
+@BreezeComponentExtension
+fun <T:Any> T.serialize(dataType:DataType):String{
+	return dataType.serialize(this)
+}
+
+/**
+ * 根据指定的数据类型，反序列化当前文本。
+ */
+@BreezeComponentExtension
+inline fun <reified T:Any> String.deserialize(dataType:DataType):T{
+	return dataType.deserialize(this,javaTypeOf<T>())
+}
+
+/**
+ * 根据指定的数据类型，反序列化当前文本。
+ */
+@BreezeComponentExtension
+fun <T:Any> String.deserialize(dataType: DataType,type:Class<T>):T{
+	return dataType.deserialize(this,type)
+}
+
+/**
+ * 根据指定的数据类型，反序列化当前文本。
+ */
+@BreezeComponentExtension
+fun <T:Any> String.deserialize(dataType: DataType,type: Type):T{
+	return dataType.deserialize(this,type)
 }
 //endregion