+++ breeze-reflect/src/main/kotlin/extensions/KTypeExtensions.kt
@@ -5,10 +5,33 @@
 
 package com.windea.breezeframework.reflect.extensions
 
+import java.lang.reflect.*
 import kotlin.reflect.*
 import kotlin.reflect.jvm.*
 
-//https://github.com/kohesive/klutter/blob/master/reflect/src/main/kotlin/uy/klutter/reflect/TypeErasure.kt
-/**得到当前Kotlin类型的被擦除类型。*/
-val KType.erasure: Class<out Any>
+/**
+ * 得到当前类型的被擦除类型。
+ */
+@Suppress("UNCHECKED_CAST") val Type.erasedType: Class<Any>
+	get() = when (this) {
+		is Class<*> -> this as Class<Any>
+		is ParameterizedType -> this.rawType.erasedType
+		is GenericArrayType -> {
+			val elementType = this.genericComponentType.erasedType
+			val testArray = java.lang.reflect.Array.newInstance(elementType, 0)
+			testArray.javaClass
+		}
+		is TypeVariable<*> -> {
+			throw IllegalStateException("Not sure what to do here yet")
+		}
+		is WildcardType -> {
+			this.upperBounds[0].erasedType
+		}
+		else -> throw IllegalStateException("Should not get here.")
+	}
+
+/**
+ * 得到当前Kotlin类型的被擦除类型。
+ */
+val KType.erasureType: Class<out Any>
 	get() = this.jvmErasure.java