+++ breeze-serialization/src/main/kotlin/KotlinxSerializer.kt
@@ -1,0 +1,12 @@
+// Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
+// Breeze is blowing...
+
+package com.windea.breezeframework.serialization
+
+/**
+ * 由Kotlinx Serialization委托实现的序列化器。
+ *
+ * @see kotlinx.serialization.SerialFormat
+ * @see kotlinx.serialization.StringFormat
+ */
+interface KotlinxSerializer : Serializer